+++ test/udp/server.js
@@ -8,6 +8,7 @@ gn.config({
 		port: 7979
 	},
 	udp: {
+		address: 'fe80::5054:1ff:fe00:3345%eth0',
 		portRange: [7980]
 		//portRange: [7980, 7990]
 	},